Refactored for separate component jsx files, restyled.
