+++ package.json
@@ -9,6 +9,7 @@
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
     "react-scripts": "4.0.3",
+    "uuidv4": "^6.2.10",
     "web-vitals": "^1.1.2"
   },
   "scripts": {
