+++ src/Game.css
@@ -1,0 +1,57 @@
+body {
+    font: 14px "Century Gothic", Futura, sans-serif;
+    text-align: center;
+    background-color: rgb(240,240,240);
+    color: rgb(88,88,88);
+}
+  
+.status {
+    margin-bottom: 10px;
+}
+  
+.square {
+    background: white;
+    border: 2px solid rgb(88,88,88);
+    float: left;
+    font-size: 48px;
+    line-height: 100px;
+    height: 100px;
+    margin-right: -1px;
+    margin-top: -1px;
+    width: 100px;
+}
+  
+.game {
+    padding: 10px;
+}
+
+.status-info {
+    margin: 10px;
+    font-size: 24px;
+}
+
+.game-board {
+    display: flex;
+    justify-content: center;
+}
+  
+.moves-info {
+    margin: 10px;
+}
+
+ol {
+    display: inline-block;
+    padding-left: 30px;
+}
+
+.move-button {
+    width: 120px;
+    background-color: rgb(88,88,88);
+    color: white;
+    border-radius: 5px;
+}
+
+li {
+    margin: 5px;
+}
+
