+++ src/Square.jsx
@@ -1,0 +1,13 @@
+import React from 'react'
+
+const Square = (props) => {
+    return (
+        <button 
+            className="square" 
+            onClick={props.onClick}
+        >
+            {props.value}
+        </button>
+    )
+}
+export default Square
